refactor(test): fold generation guard test into test_delete_session_bridge_cleanup
Move ``test_spec_fill_parked_across_delete_discards`` (and its
``_AgentSnapshotServerClient`` helper) from the now-deleted
``tests/runner/test_session_cache_generation.py`` into
``tests/runner/test_delete_session_bridge_cleanup.py``, which already
covers what ``delete_session`` cleans up.  Update the module docstring
to reflect both issues the file now guards against (#1350 bridge-dir
leak, #3728 ABA spec-cache repopulation).
